Incorporating @ptrovatelli's suggestions.
